+++ src/test/java/com/formento/neighborhood/model/NodeTest.java
@@ -27,7 +27,7 @@ public class NodeTest {
     private Node root;
 
     private Property mapFromPoint(Point point) {
-        return new Property(Optional.empty(), "title", 123, "description", point, (short) 10, (short) 11, 321);
+        return new Property(Optional.empty(), "title", 123, "description", point,  10,  11, 321);
     }
 
     /* tree
